Se han creado las dos entidades que van a interactuar de momento, el usuario y el computador, procederemos a crear la vista para hacer mas funcional el app.

+++ computer_store/backend/src/main/java/com/computer/computer/DTO/User/CreateUserDTO.java
@@ -15,9 +15,9 @@ public class CreateUserDTO {
 
     private String lastname;
 
-    private int identificationDocument;
+    private Long identificationDocument;
 
-    private int numberPhone;
+    private Long numberPhone;
 
     private String email;
